service/auth-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
cryptography==41.0.7
psycopg2-binary==2.9.9
email-validator==2.1.0
pandas==2.2.1
openpyxl==3.1.2
